refact: separate itemsRouter
